Removed --list, replaced by --parseable. Outputs plain text results, no titles, no footer. Uses colors
